backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-api-python-client==2.97.0
python-dotenv==1.0.0 